ECOM APP CHANGES
ECOM APP CHANGES

+++ inventory_service/src/main/java/com/shoppix/inventory_service_reactive/events/InventoryEvent.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class InventoryEvent {
 
-    private int inventoryId;
+    private long inventoryId;
 
     private String parentProductId;
